requirements.txt
aiokafka~=0.11.0
alembic~=1.13.1
asyncpg~=0.29.0
authlib~=1.3.0
bcrypt~=3.1.7
click~=8.1.7
elastic-apm~=6.23.0
fastapi~=0.111.0
fastapi_pagination
httpx~=0.27.0
httptools~=0.6.1
itsdangerous
passlib~=1.7.4
phonenumbers~=8.13.36
psycopg2-binary
pydantic~=2.7.1
python-dotenv~=1.0.1
python-jose~=3.3
pip~=24.2
requests~=2.32.2
SQLAlchemy~=2.0.30
starlette~=0.37.2
uvicorn==0.30.5
uvloop==0.20.0

fastapi-pagination~=0.12.31
Jinja2~=3.1.4
beautifulsoup4==4.12.3